requirements.txt
crewai==0.28.8
streamlit
groq
langchain-groq
python-dotenv
datetime